+++ sources/packages/web/src/services/http/dto/Config.dto.ts
@@ -1,0 +1,12 @@
+import { ClientIdType } from "@/types";
+
+export interface AuthConfigAPIOutDTO {
+  url: string;
+  realm: string;
+  clientIds: { [Value in ClientIdType]: string };
+  externalSiteMinderLogoutUrl?: string;
+}
+
+export interface ConfigAPIOutDTO {
+  auth: AuthConfigAPIOutDTO;
+}
