+++ sources/packages/web/src/services/http/dto/index.ts
@@ -19,3 +19,5 @@ export * from "@/services/http/dto/Restriction.dto";
 export * from "@/services/http/dto/Application.dto";
 export * from "@/services/http/dto/ProgramYear.dto";
 export * from "@/services/http/dto/User.dto";
+export * from "@/services/http/dto/Config.dto";
+export * from "@/services/http/dto/DesignationAgreement.dto";
