+++ sources/packages/web/src/views/aest/institution/DesignationAESTView.vue
@@ -64,7 +64,7 @@ import {
   DesignationAgreementStatus,
   DesignationLocationAPIInDTO,
   DesignationAgreementAPIOutDTO,
-} from "@/types/contracts/DesignationAgreementContract";
+} from "@/services/http/dto";
 import {
   DesignationModel,
   DesignationFormViewModes,
